requirements.txt
amqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.0.1
billiard==3.6.4.0
boto3==1.26.142
botocore==1.29.142
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
decorator==5.1.1
Django==4.2.1
django-cors-headers==4.0.0
django-crontab==0.7.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
fabric==3.1.0
idna==3.4
invoke==2.1.2
jmespath==1.0.1
kombu==5.2.4
paramiko==3.2.0
prompt-toolkit==3.0.38
pycparser==2.21
pyignite==0.6.1
pymemcache==4.0.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python3-memcached==1.51
pytz==2023.3
redis==4.5.5
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
